+++ modules/operations/src/main/java/com/yrootlab/onmaru/operations/admission/SubjectType.java
@@ -1,6 +1,7 @@
 package com.yrootlab.onmaru.operations.admission;
 
 public enum SubjectType {
+    GUEST,
     IP,
     MEMBER
 }